+++ ta4j-examples/src/main/java/ta4jexamples/loaders/CsvBarsLoader.java
@@ -27,19 +27,20 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
+import java.time.Instant;
 import java.time.LocalDate;
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
+import java.time.ZoneOffset;
 import java.time.format.DateTimeFormatter;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+
+import org.ta4j.core.BarSeries;
+import org.ta4j.core.BaseBarSeriesBuilder;
 
 import com.opencsv.CSVParserBuilder;
 import com.opencsv.CSVReader;
 import com.opencsv.CSVReaderBuilder;
 import com.opencsv.exceptions.CsvValidationException;
-import org.ta4j.core.BarSeries;
-import org.ta4j.core.BaseBarSeriesBuilder;
 
 /**
  * This class build a Ta4j bar series from a CSV file containing bars.
@@ -70,7 +71,7 @@ public class CsvBarsLoader {
                     .build()) {
                 String[] line;
                 while ((line = csvReader.readNext()) != null) {
-                    ZonedDateTime date = LocalDate.parse(line[0], DATE_FORMAT).atStartOfDay(ZoneId.systemDefault());
+                    Instant date = LocalDate.parse(line[0], DATE_FORMAT).atStartOfDay(ZoneOffset.UTC).toInstant();
                     double open = Double.parseDouble(line[1]);
                     double high = Double.parseDouble(line[2]);
                     double low = Double.parseDouble(line[3]);